Extract shared LEVEL_COLORS and formatTime into lib/format.ts
Deduplicates identical LEVEL_COLORS definitions from 4 components and
formatTime from 2 components into a single shared module.

Closes #33
