+++ src/components/LogTable.tsx
@@ -1,6 +1,7 @@
 import {Text, Box, useInput} from 'ink';
 import {useState} from 'react';
 import type {LogEntry, LogsResponse} from '../types/log.js';
+import {LEVEL_COLORS, formatTime} from '../lib/format.js';
 import LogDetail from './LogDetail.js';
 
 type Props = {
@@ -8,22 +9,6 @@ type Props = {
 	pagination: LogsResponse['pagination'];
 	filterSummary?: string;
 };
-
-const LEVEL_COLORS: Record<string, string> = {
-	debug: 'gray',
-	info: 'blue',
-	warn: 'yellow',
-	error: 'red',
-};
-
-function formatTime(timestamp: string): string {
-	const d = new Date(timestamp);
-	const h = String(d.getHours()).padStart(2, '0');
-	const m = String(d.getMinutes()).padStart(2, '0');
-	const s = String(d.getSeconds()).padStart(2, '0');
-	const ms = String(d.getMilliseconds()).padStart(3, '0');
-	return `${h}:${m}:${s}.${ms}`;
-}
 
 export default function LogTable({logs, pagination, filterSummary}: Props) {
 	const [cursor, setCursor] = useState(0);
